dataplatform/projects/Komfyrvakt/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
redis>=5.0.0
python-dateutil>=2.8.2
google-genai>=0.2.0
python-dotenv>=1.0.0

